% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krr_fit.R
\name{gaussian_kernel}
\alias{gaussian_kernel}
\title{Gaussian kernel function}
\usage{
gaussian_kernel(x, z, rho = 1)
}
\arguments{
\item{x}{A numeric vector.}

\item{z}{A numeric vector of the same length as x.}

\item{rho}{Kernel hyper-parameter (default = 1).}
}
\value{
A numeric scalar representing the kernel value.
}
\description{
Computes the Gaussian kernel value between two vectors x and z.
}
\examples{
gaussian_kernel(c(1,2), c(1,2))
gaussian_kernel(c(1,0), c(0,1), rho = 0.5)
}
